refactoring MainActivity
