+++ app/src/main/AndroidManifest.xml
@@ -9,8 +9,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".NewPageActivity" />
-        <activity android:name=".PageDetailActivity" />
+        <activity android:name=".ui.NewPageActivity" />
+        <activity android:name=".ui.PageDetailActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
