+++ app/src/main/java/hidaka/kotlinstudy/todo/ui/NewPageActivity.kt
@@ -1,12 +1,13 @@
-package hidaka.kotlinstudy.todo
+package hidaka.kotlinstudy.todo.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import hidaka.kotlinstudy.todo.R
 
-class PageDetailActivity : AppCompatActivity() {
+class NewPageActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_page_detail)
+        setContentView(R.layout.activity_new_page)
     }
 }
