+++ app/src/main/java/hidaka/kotlinstudy/todo/ui/PageDetailActivity.kt
@@ -1,12 +1,13 @@
-package hidaka.kotlinstudy.todo
+package hidaka.kotlinstudy.todo.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import hidaka.kotlinstudy.todo.R
 
-class NewPageActivity : AppCompatActivity() {
+class PageDetailActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_new_page)
+        setContentView(R.layout.activity_page_detail)
     }
 }
