+++ app/src/main/res/layout/activity_page_detail.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PageDetailActivity">
+    tools:context=".ui.PageDetailActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
